requirements.txt
# Essential packages
setuptools>=65.0.0
wheel

# Django Framework
Django==5.2.2
asgiref==3.8.1
sqlparse==0.5.0
tzdata==2024.1

# Django REST Framework and JWT
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.27.2
PyYAML==6.0.1
uritemplate==4.1.1

# CORS Headers
django-cors-headers==4.4.0

# Database
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.7
django-redis==5.4.0

# Celery for async tasks
celery==5.3.4
kombu==5.3.4
billiard==4.2.0

# Social Authentication
social-auth-app-django==5.4.2
social-auth-core==4.5.4

# Djoser for authentication
djoser==2.2.3

# Cloud Storage (Cloudinary)
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Image Processing
Pillow==10.4.0

# Environment Configuration
python-decouple==3.8

# Filters
django-filter==24.3

# Google OAuth
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Stripe for payments
stripe==10.8.0

# HTTP Requests
requests==2.32.3
urllib3==2.2.2

# Cryptography
cryptography==43.0.0

# Other utilities
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
six==1.16.0
pytz==2024.1

# Development dependencies (optional)
# Uncomment for development
# django-debug-toolbar==4.4.6
# ipython==8.26.0
